Overhaul of workflow document structure, so instead of putting everything in one document, we now have nicely sorted FireTasks, Fireworks and Workflows
